Reorgnised align_up from Heap.cpp to types.h

+++ src/kernel/include/mapple/scheduling/Proccess.h
@@ -5,7 +5,7 @@
 class Proccess{
     public:
     ll_t node;
-    uint64_t pid;
+    uint8_t pid;
     uint64_t importance;
     uint8_t isWaiting;
     void wait();
